multi: move URL{De,En}coder functions to asset package
To avoid circular package dependencies in the next commits, we move the
URLDecoder and URLEncoder functions from the address to the asset
package. That way we'll be able to also use it from the proof package.
