+++ address/encoding.go
@@ -2,7 +2,6 @@ package address
 
 import (
 	"io"
-	"net/url"
 
 	"github.com/btcsuite/btcd/btcec/v2"
 	"github.com/lightningnetwork/lnd/tlv"
@@ -40,35 +39,6 @@ func compressedPubKeyDecoder(r io.Reader, val any, buf *[8]byte, l uint64) error
 	return tlv.NewTypeForDecodingErr(
 		val, "*btcec.PublicKey", l, btcec.PubKeyBytesLenCompressed,
 	)
-}
-
-// UrlEncoder encodes a url.URL as a variable length byte slice.
-func UrlEncoder(w io.Writer, val any, buf *[8]byte) error {
-	if t, ok := val.(*url.URL); ok {
-		addrBytes := []byte((*t).String())
-		return tlv.EVarBytes(w, &addrBytes, buf)
-	}
-	return tlv.NewTypeForEncodingErr(val, "*url.URL")
-}
-
-// UrlDecoder decodes a variable length byte slice as an url.URL.
-func UrlDecoder(r io.Reader, val any, buf *[8]byte, l uint64) error {
-	if t, ok := val.(*url.URL); ok {
-		var addrBytes []byte
-		err := tlv.DVarBytes(r, &addrBytes, buf, l)
-		if err != nil {
-			return err
-		}
-
-		addr, err := url.ParseRequestURI(string(addrBytes))
-		if err != nil {
-			return err
-		}
-		*t = *addr
-
-		return nil
-	}
-	return tlv.NewTypeForDecodingErr(val, "*url.URL", l, l)
 }
 
 func VersionEncoder(w io.Writer, val any, buf *[8]byte) error {
